+++ anonymousletter/src/main/java/com/somanyteam/event/dto/result/user/UserGetInfoResult.java
@@ -19,7 +19,7 @@ public class UserGetInfoResult {
     private String email;
 
     @ApiModelProperty(value = "性别,0-男 1-女 可不填", required = true)
-    private Integer sex;
+    private Byte sex;
 
     @JsonFormat(pattern = "yyyy-MM-dd",timezone="GMT+8")
     @DateTimeFormat(pattern = "yyyy-MM-dd")
